pyproject.toml
[project]
name = "django-velocity"
version = "0.1.0"
description = "Modern Django Boilerplate with Service-Oriented Architecture"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "you@example.com" }]

dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "django-ninja>=1.3",
    "djangorestframework-simplejwt>=5.3",
    "psycopg[binary]>=3.2",
    "django-environ>=0.11",
    "whitenoise>=6.7",
    "python-dateutil>=2.9.0.post0",
    "django-allauth[mfa,socialaccount]>=65.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-cov>=4.1",
    "factory-boy>=3.3",
    "ruff>=0.5",
    "pre-commit>=3.7",
    "ipython>=8.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-cov>=4.1",
    "factory-boy>=3.3",
    "ruff>=0.5",
    "pre-commit>=3.7",
    "ipython>=8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["apps", "config"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "DJ",     # flake8-django
    "SIM",    # flake8-simplify
    "PTH",    # flake8-use-pathlib
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.django.test"
python_files = ["test_*.py", "*_test.py"]
addopts = [
    "--strict-markers",
    "-ra",
]

[tool.coverage.run]
source = ["apps"]
branch = true
omit = ["*/migrations/*", "*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
